Backend/requirements.txt
fastapi
uvicorn
requests
python-dotenv
pydantic
langchain
langchain-openai
openai
scikit-learn
